modification de l'id de chaque objet de integer en long

+++ Poseiden-skeleton/src/main/java/com/poseidon/api/controllers/BidController.java
@@ -53,7 +53,7 @@ public class BidController {
     }
 
     @GetMapping("/bidList/update/{id}")
-    public String showUpdateForm(@PathVariable("id") Integer id, Model model, RedirectAttributes redirectAttributes) {
+    public String showUpdateForm(@PathVariable("id") Long id, Model model, RedirectAttributes redirectAttributes) {
         Bid bidToUpdate = bidService.findBidById(id);
         BidDto bidDto = bidService.convertEntityToDto(bidToUpdate);
         model.addAttribute("bidDto", bidDto);
@@ -62,7 +62,7 @@ public class BidController {
     }
 
     @PostMapping("/bidList/update/{id}")
-    public String updateBid(@PathVariable("id") Integer id, @Valid BidDto bidDto,
+    public String updateBid(@PathVariable("id") Long id, @Valid BidDto bidDto,
                             BindingResult result, Model model, RedirectAttributes redirectAttributes) {
 
         if (result.hasErrors()) {
@@ -78,7 +78,7 @@ public class BidController {
     }
 
     @GetMapping("/bidList/delete/{id}")
-    public String deleteBid(@PathVariable("id") Integer id, Model model, RedirectAttributes redirectAttributes) {
+    public String deleteBid(@PathVariable("id") Long id, Model model, RedirectAttributes redirectAttributes) {
         bidService.deleteBid(id);
         redirectAttributes.addFlashAttribute("message", String.format("Bid with id " + id + " was successfully deleted"));
         model.addAttribute("bids", bidService.findAllBids());
