+++ Poseiden-skeleton/src/main/java/com/poseidon/api/repositories/TradeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TradeRepository extends JpaRepository<Trade, Integer> {
+public interface TradeRepository extends JpaRepository<Trade, Long> {
 }
