+++ Poseiden-skeleton/src/main/java/com/poseidon/api/repositories/UserRepository.java
@@ -9,10 +9,10 @@ import java.util.Optional;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
 
     Optional<User> findByUsername(String username);
 
     @Override
-    Optional<User> findById(Integer userId);
+    Optional<User> findById(Long userId);
 }
